init webpack.config.ts
