+++ package.json
@@ -1,12 +1,12 @@
 {
   "name": "node.js",
   "version": "0.1.0",
-  "description": "Make CRUD with Node.js backend",
+  "description": "Make CRUD with Node.js, Express backend and MongoDB for database",
   "main": "index.js",
   "scripts": {
-    "start:server": "nodemon ./dist/index.js",
-    "watch": "tsc -w -p ./tsconfig.json",
+    "start:server": "nodemon ./dist/server.js",
     "compile": "tsc -p ./tsconfig.json",
+    "build:webpack-dev": "webpack",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -24,16 +24,23 @@
     "axios": "^0.19.2",
     "cors": "^2.8.5",
     "express": "^4.17.1",
-    "firebase-admin": "^8.10.0",
+    "mongoose": "^5.9.9",
     "xml-js": "^1.6.11"
   },
   "devDependencies": {
     "@types/body-parser": "^1.19.0",
     "@types/cors": "^2.8.6",
-    "@types/express": "^4.17.3",
-    "@types/node": "^13.9.2",
+    "@types/express": "^4.17.6",
+    "@types/mongoose": "^5.7.11",
+    "@types/node": "^13.11.1",
+    "@types/webpack-node-externals": "^1.7.1",
     "body-parser": "^1.19.0",
-    "nodemon": "^2.0.2",
-    "typescript": "^3.8.3"
+    "nodemon": "^2.0.3",
+    "ts-loader": "^7.0.0",
+    "ts-node": "^8.8.2",
+    "typescript": "^3.8.3",
+    "webpack": "^4.42.1",
+    "webpack-cli": "^3.3.11",
+    "webpack-node-externals": "^1.7.2"
   }
 }
