+++ tsconfig.json
@@ -1,8 +1,8 @@
 {
   "compilerOptions": {
     "outDir": "./dist",
-    "module": "commonjs",
-    "target": "ES2017",
+    "module": "CommonJS",
+    "target": "ES2020",
     "strict": true,
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
@@ -10,11 +10,11 @@
     "removeComments": true,
     "moduleResolution": "node"
   },
-  "file": [
-  ],
   "exclude": [
     "node_modules",
-    "src/*/*.ts",
     "frontend"
+  ],
+  "include": [
+    "src/**/*.ts"
   ]
 }
